refactor(framework): Move grpc module
