+++ framework/py/flwr/client/grpc_adapter_client/connection.py
@@ -24,10 +24,10 @@ from cryptography.hazmat.primitives.asymmetric import ec
 from flwr.app.message import Message
 from flwr.client.grpc_rere_client.connection import grpc_request_response
 from flwr.client.grpc_rere_client.grpc_adapter import GrpcAdapter
-from flwr.common import GRPC_MAX_MESSAGE_LENGTH
 from flwr.common.logger import log
 from flwr.proto.message_pb2 import ObjectTree  # pylint: disable=E0611
 from flwr.supercore.fab import Fab
+from flwr.supercore.grpc import GRPC_MAX_MESSAGE_LENGTH
 from flwr.supercore.retry import RetryInvoker
 from flwr.supercore.run import Run
