+++ framework/py/flwr/client/grpc_rere_client/node_auth_client_interceptor_test.py
@@ -28,7 +28,6 @@ from parameterized import parameterized
 
 from flwr.app.message import Message, RecordDict
 from flwr.client.grpc_rere_client.connection import grpc_request_response
-from flwr.common import GRPC_MAX_MESSAGE_LENGTH
 from flwr.common.constant import PUBLIC_KEY_HEADER, SIGNATURE_HEADER, TIMESTAMP_HEADER
 from flwr.proto.fleet_pb2 import (  # pylint: disable=E0611
     ActivateNodeRequest,
@@ -54,6 +53,7 @@ from flwr.supercore.constant import (
     FLWR_PACKAGE_NAME_METADATA_KEY,
     FLWR_PACKAGE_VERSION_METADATA_KEY,
 )
+from flwr.supercore.grpc import GRPC_MAX_MESSAGE_LENGTH
 from flwr.supercore.primitives.asymmetric import (
     generate_key_pairs,
     public_key_to_bytes,
