+++ framework/py/flwr/compat/client/app.py
@@ -33,7 +33,6 @@ from flwr.client.message_handler.message_handler import handle_control_message
 from flwr.client.numpy_client import NumPyClient
 from flwr.client.run_info_store import DeprecatedRunInfoStore
 from flwr.clientapp.client_app import ClientApp, LoadClientAppError
-from flwr.common import GRPC_MAX_MESSAGE_LENGTH
 from flwr.common.constant import MAX_RETRY_DELAY, ErrorCode
 from flwr.common.exit import ExitCode, flwr_exit
 from flwr.common.logger import log, warn_deprecated_feature
@@ -42,6 +41,7 @@ from flwr.compat.client.typing import ClientFnExt
 from flwr.compat.common.constant import TRANSPORT_TYPE_GRPC_BIDI, TRANSPORT_TYPES_COMPAT
 from flwr.supercore.address import parse_address
 from flwr.supercore.fab import Fab
+from flwr.supercore.grpc import GRPC_MAX_MESSAGE_LENGTH
 from flwr.supercore.object_store import ObjectStoreFactory
 from flwr.supercore.retry import RetryInvoker, RetryState, exponential
 from flwr.supercore.run import Run, RunNotRunningException
