+++ framework/py/flwr/server/app.py
@@ -30,7 +30,6 @@ from typing import TypeVar, cast
 import grpc
 import yaml
 
-from flwr.common import GRPC_MAX_MESSAGE_LENGTH
 from flwr.common.args import (
     add_args_runtime_dependency_install,
     try_obtain_server_certificates,
@@ -55,7 +54,6 @@ from flwr.common.constant import (
 )
 from flwr.common.event_log_plugin import EventLogWriterPlugin
 from flwr.common.exit import ExitCode, flwr_exit, register_signal_handlers
-from flwr.common.grpc import generic_create_grpc_server
 from flwr.common.logger import configure_superlink_log_file, log
 from flwr.proto.fleet_pb2_grpc import (  # pylint: disable=E0611
     add_FleetServicer_to_server,
@@ -68,6 +66,7 @@ from flwr.supercore.auth import (
     load_superexec_auth_secret,
 )
 from flwr.supercore.constant import FLWR_IN_MEMORY_DB_NAME
+from flwr.supercore.grpc import GRPC_MAX_MESSAGE_LENGTH, generic_create_grpc_server
 from flwr.supercore.grpc_health import add_args_health, run_health_server_grpc_no_tls
 from flwr.supercore.interceptors import create_fleet_runtime_version_server_interceptor
 from flwr.supercore.object_store import ObjectStoreFactory
