+++ framework/py/flwr/supercore/grpc.py
@@ -24,9 +24,8 @@ from typing import Any
 
 import grpc
 
+from flwr.common.logger import log
 from flwr.supercore.address import is_port_in_use
-
-from .logger import log
 
 GRPC_MAX_MESSAGE_LENGTH: int = 2_147_483_647  # == 2048 * 1024 * 1024 -1 (2GB)
